feat: Add ngram repeat block on kernel_fusion_utils
